KAFKA-14209: Change Topology optimization to accept list of rules 1/3 (#12641) (#6)
This PR is part of a series implementing the self-join rewriting. As part of it, we decided to clean up the TOPOLOGY_OPTIMIZATION_CONFIG and make it a list of optimization rules. Acceptable values are: NO_OPTIMIZATION, OPTIMIZE which applies all optimization rules or a comma separated list of specific optimizations.

Reviewers: Guozhang Wang <guozhang@apache.org>, John Roesler <vvcephei@apache.org>

Co-authored-by: Vicky Papavasileiou <vpapavas@users.noreply.github.com>